Check if tables exist

+++ db/migrate/20140105194519_move_admin_notes_to_comments.rb
@@ -1,17 +1,19 @@
 class MoveAdminNotesToComments < ActiveRecord::Migration
   def self.up
-    remove_index  :admin_notes, [:admin_user_type, :admin_user_id]
-    rename_table  :admin_notes, :active_admin_comments
-    rename_column :active_admin_comments, :admin_user_type, :author_type
-    rename_column :active_admin_comments, :admin_user_id, :author_id
-    add_column    :active_admin_comments, :namespace, :string
-    add_index     :active_admin_comments, [:namespace]
-    add_index     :active_admin_comments, [:author_type, :author_id]
+    unless ActiveRecord::Base.connection.table_exists? 'active_admin_comments'
+      remove_index  :admin_notes, [:admin_user_type, :admin_user_id]
+      rename_table  :admin_notes, :active_admin_comments
+      rename_column :active_admin_comments, :admin_user_type, :author_type
+      rename_column :active_admin_comments, :admin_user_id, :author_id
+      add_column    :active_admin_comments, :namespace, :string
+      add_index     :active_admin_comments, [:namespace]
+      add_index     :active_admin_comments, [:author_type, :author_id]
 
-    # Update all the existing comments to the default namespace
-    say "Updating any existing comments to the #{ActiveAdmin.application.default_namespace} namespace."
-    comments_table_name = ActiveRecord::Migrator.proper_table_name("active_admin_comments")
-    execute "UPDATE #{comments_table_name} SET namespace='#{ActiveAdmin.application.default_namespace}'"
+      # Update all the existing comments to the default namespace
+      say "Updating any existing comments to the #{ActiveAdmin.application.default_namespace} namespace."
+      comments_table_name = ActiveRecord::Migrator.proper_table_name("active_admin_comments")
+      execute "UPDATE #{comments_table_name} SET namespace='#{ActiveAdmin.application.default_namespace}'"
+    end
   end
 
   def self.down
